Split entity tracking into its own module.
